Move shared mocks into a shared helper file
