+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/ScatterPlotViewer/components/Axes/Axes.js
@@ -2,11 +2,12 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { Group } from '@vx/group'
 import Axis from './components/Axis'
+import { MARGIN, PADDING } from '../../helpers/constants'
 
 function Axes ({ axesConfig, ...rest }) {
   const { xAxis, yAxis } = axesConfig
   return (
-    <Group>
+    <Group left={MARGIN}>
       <Axis axis={yAxis} {...rest} />
       <Axis axis={xAxis} {...rest} />
     </Group>
