+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/ScatterPlotViewer/components/Axes/components/Axis/Axis.spec.js
@@ -1,54 +1,20 @@
-import { mount, shallow } from 'enzyme'
+import { shallow } from 'enzyme'
 import React from 'react'
-import * as d3 from 'd3'
-import { scaleLinear } from '@vx/scale'
-import { genRandomNormalPoints } from '@vx/mock-data'
 import { AxisLeft, AxisBottom } from '@vx/axis'
 import zooTheme from '@zooniverse/grommet-theme'
 import { Axis } from './Axis'
 import InnerTickAxis from '../InnerTickAxis'
-import { MARGIN, PADDING } from '../../../../helpers/constants'
-
-
-const parentWidth = 768
-const parentHeight = 384
-const color = zooTheme.global.colors['light-1']
-const fontFamily = zooTheme.global.font.family
-const fontSize = zooTheme.text.xsmall.size
-
-const randomPoints = genRandomNormalPoints()
-const xPoints = randomPoints.filter((point) => {
-  return point[0]
-})
-const yPoints = randomPoints.filter((point) => {
-  return point[1]
-})
-const dataExtent = {
-  x: d3.extent(xPoints),
-  y: d3.extent(yPoints)
-}
-
-const xScale = scaleLinear({
-  domain: dataExtent.x,
-  range: [0 + PADDING, parentWidth - MARGIN]
-})
-
-const yScale = scaleLinear({
-  domain: dataExtent.y,
-  range: [parentHeight - PADDING, 0 + MARGIN]
-})
-
-const bottomAxis = {
-  label: 'Days',
-  orientation: 'bottom',
-  scale: xScale
-}
-
-const leftAxis = {
-  label: 'Brightness',
-  orientation: 'left',
-  scale: yScale
-}
+import { PADDING } from '../../../../helpers/constants'
+import {
+  bottomAxis,
+  color,
+  fontFamily,
+  fontSize,
+  leftAxis,
+  parentWidth,
+  parentHeight,
+  xScale
+} from '../../../../helpers/mockData'
 
 describe('Component > Axis', function () {
   it('should render without crashing', function () {
