+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/ScatterPlotViewer/components/Axes/components/InnerTickAxis/InnerTickAxis.spec.js
@@ -1,54 +1,19 @@
-import { mount, shallow } from 'enzyme'
+import { mount } from 'enzyme'
 import React from 'react'
-import * as d3 from 'd3'
-import { scaleLinear } from '@vx/scale'
-import { genRandomNormalPoints } from '@vx/mock-data'
 import { Axis } from '@vx/axis'
-import { Group } from '@vx/group'
 import { Line } from '@vx/shape'
 import zooTheme from '@zooniverse/grommet-theme'
 import InnerTickAxis from './InnerTickAxis'
 import { MARGIN, PADDING } from '../../../../helpers/constants'
-
-const parentWidth = 768
-const parentHeight = 384
-const color = zooTheme.global.colors['light-1']
-const fontSize = zooTheme.text.xsmall.size
-
-const randomPoints = genRandomNormalPoints()
-const xPoints = randomPoints.map((point) => {
-  return point[0]
-})
-const yPoints = randomPoints.map((point) => {
-  return point[1]
-})
-
-const dataExtent = {
-  x: d3.extent(xPoints),
-  y: d3.extent(yPoints)
-}
-
-const xScale = scaleLinear({
-  domain: dataExtent.x,
-  range: [0 + PADDING, parentWidth - MARGIN]
-})
-
-const yScale = scaleLinear({
-  domain: dataExtent.y,
-  range: [parentHeight - PADDING, 0 + MARGIN]
-})
-
-const bottomAxis = {
-  label: 'Days',
-  orientation: 'bottom',
-  scale: xScale
-}
-
-const leftAxis = {
-  label: 'Brightness',
-  orientation: 'left',
-  scale: yScale
-}
+import {
+  bottomAxis,
+  color,
+  fontSize,
+  leftAxis,
+  parentWidth,
+  parentHeight,
+  xScale
+} from '../../../../helpers/mockData'
 
 describe('Component > InnerTickAxis', function () {
   describe('render', function () {
